fix: character encoding on Update all #87
