+++ server/routes/api/utils/index.js
@@ -1,10 +1,8 @@
 import { parseBookmarkFile, parseHtml } from "../../../helpers/parsers.js";
-import { batchUpdateLinks } from "../../../logics/bookmarks.js";
-
-import axios from "axios";
-
-const DEFAULT_USER_AGENT =
-  "Mozilla/5.0 (Windows; Windows NT 10.4; Win64; x64) AppleWebKit/535.50 (KHTML, like Gecko) Chrome/48.0.3463.124 Safari/536";
+import {
+  batchUpdateLinks,
+  requestHeadFromUrl,
+} from "../../../logics/bookmarks.js";
 
 /**
  *
@@ -29,61 +27,8 @@ export default async function (fastify, opts) {
     },
     async function (request, reply) {
       let { url } = request.body;
-      let res;
-      const controller = new AbortController();
-
-      try {
-        res = await axios.get(url, {
-          responseType: "stream",
-          signal: controller.signal,
-          headers: {
-            Accept: "text/html,application/xhtml+xml,application/xml",
-            Dnt: "1",
-            "Upgrade-Insecure-Requests": "1",
-            "User-Agent": DEFAULT_USER_AGENT,
-          },
-        });
-      } catch (error) {
-        console.error(url, error.message);
-        res = error.response;
-      }
-
-      let contentType = res.headers["content-type"];
-      let matches = contentType.match(/charset=\s*"?(.[^\"]+)"?$/i);
-      let encoding = matches?.[1] || "utf-8";
-
-      let stream = res.data;
-
-      stream.setEncoding("hex");
-
-      let htmlArray = [];
-
-      stream.on("data", (chunk) => {
-        let chunkHTML = new TextDecoder(encoding.toLowerCase()).decode(
-          Buffer.from(chunk, "hex")
-        );
-        const closingHeadTag = "</head>";
-        const closingHeadTagIndex =
-          chunkHTML.indexOf(closingHeadTag) + closingHeadTag.length;
-
-        if (chunkHTML.includes(closingHeadTag)) {
-          chunkHTML = chunkHTML.substring(0, closingHeadTagIndex);
-          htmlArray.push(chunkHTML);
-          controller.abort();
-          return;
-        }
-        htmlArray.push(chunkHTML);
-      });
-
-      return new Promise((resolve, reject) => {
-        stream.on("close", () => {
-          const htmlOut = htmlArray.join("").trim();
-          resolve(parseHtml(htmlOut, url));
-        });
-        stream.on("error", (e) => {
-          if (e.code !== "ERR_CANCELED") reject(e);
-        });
-      });
+      let headHtml = await requestHeadFromUrl(url);
+      return parseHtml(headHtml, url);
     }
   );
